LayoutTests/fast/animation/request-animation-frame-timestamps-advance-expected.txt

<<<<<<< HEAD
Tests the timestamps provided to requestAnimationFrame callbacks advance

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstTimestamp is defined.
PASS secondTimestamp is defined.
PASS secondTimestamp > firstTimestamp is true
PASS successfullyParsed is true

TEST COMPLETE
=======
Tests the timestamps provided to requestAnimationFrame callbacks advance

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstTimestamp is defined.
PASS secondTimestamp is defined.
PASS secondTimestamp > firstTimestamp is true
PASS successfullyParsed is true

TEST COMPLETE
>>>>>>> 74e48da1f23520600d05ce61eb732f6bd2dc87af
